pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "pandas>=2.3.1",
    "streamlit-folium>=0.25.0",
    "streamlit>=1.46.1",
    "plotly>=6.2.0",
    "numpy>=2.3.1",
    "folium>=0.20.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "scikit-learn>=1.7.0",
]
